Fixes to tests after restructuring code to support npm and bower packaging.
